fix: change qemu build to oriole
The array we grab the version from was re-ordered as PG17 non-oriole
was introduced. A more robust mechanism can be added later but this
array is rarely adjusted, so this is likely good enough for now.

Also moves up some fixes from the init-scripts, and removes Kong from
the qemu artifact.